+++ notify-impl/src/test/java/com/dwarfeng/notify/impl/service/SenderSenderSenderRelationMaintainServiceImplTest.java
@@ -43,7 +43,7 @@ public class SenderSenderSenderRelationMaintainServiceImplTest {
     @Before
     public void setUp() {
         notifySetting = new NotifySetting(null, "label", "remark");
-        topic = new Topic(new StringIdKey("test.topic"), "label", "remark");
+        topic = new Topic(new StringIdKey("test.topic"), "label", "remark", true, 12450);
         senderInfo = new SenderInfo(null, "label", "type", "param", "remark");
         senderRelation = new SenderRelation(null, null, "remark");
     }